backend_FastApi/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
ultralytics==8.3.24
google-cloud-storage==2.18.2

